requirements.txt
boto3==1.26.137
aws-iot-device-sdk-python-v2==1.7.0
paho-mqtt==1.6.1
numpy==1.24.3
pandas==2.0.3
pytest==7.3.1
python-dotenv==1.0.0
requests==2.31.0
pyserial==3.5
scipy==1.10.1
matplotlib==3.7.1 